+++ packages/blade/src/utils/getFocusRingStyles/types.ts
@@ -3,6 +3,7 @@ import type { Theme } from '~components/BladeProvider';
 type GetFocusRingArgs = {
   theme: Theme;
   negativeOffset?: boolean;
+  isImportant?: boolean;
 };
 
 export type { GetFocusRingArgs };